requirements.txt
# Pneumonia Detection Challenge - Project Dependencies

# Core Data Science Libraries
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Deep Learning
tensorflow>=2.14.0
keras>=2.14.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# Web Application
streamlit>=1.28.0

# Dataset Management
kagglehub>=0.2.4

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
